+++ knble/src/main/java/ovh/karewan/knble/ble/DeviceOp.java
@@ -59,14 +59,15 @@ public class DeviceOp {
 	// Always use main thread with BluetoothGatt to avoid issues
 	private final Handler mMainHandler = new Handler(Looper.getMainLooper());
 	private final BleDevice mDevice;
-	private BluetoothGatt mBluetoothGatt;
-	private BleGattCallback mCallback;
-	private int mState = BleGattCallback.DISCONNECTED;
-	private int lastGattStatus = 0;
-	private boolean lastWriteSuccess = false;
 
+	private BleGattCallback mCallback;
+
+	private int mState = BleGattCallback.DISCONNECTED;
+	private BluetoothGatt mBluetoothGatt;
+	private int mLastGattStatus = 0;
+
+	private boolean mLastWriteSuccess = false;
 	private BleWriteCallback mWriteCallback;
-	private BluetoothGattService mWriteService;
 	private BluetoothGattCharacteristic mWriteCharacteristic;
 	private Queue<byte[]> mWriteQueue;
 	private int mWriteTotalPkg;
@@ -74,7 +75,6 @@ public class DeviceOp {
 	private long mWriteInterval;
 
 	private BleReadCallback mReadCallback;
-	private BluetoothGattService mReadService;
 	private BluetoothGattCharacteristic mReadCharacteristic;
 
 	/**
@@ -95,20 +95,20 @@ public class DeviceOp {
 	}
 
 	/**
-	 * Get the handler
-	 * @return mMainHandler
-	 */
-	public Handler getHandler() {
-		return mMainHandler;
-	}
-
-	/**
 	 * Get the bluetooth gatt
 	 * @return mBluetoothGatt
 	 */
 	@Nullable
 	public BluetoothGatt getBluetoothGatt() {
 		return mBluetoothGatt;
+	}
+
+	/**
+	 * Set current connection state
+	 * @param state state
+	 */
+	private synchronized void setState(int state) {
+		mState = state;
 	}
 
 	/**
@@ -129,10 +129,82 @@ public class DeviceOp {
 
 	/**
 	 * Get the last gatt status
-	 * @return lastGattStatus
+	 * @return mLastGattStatus
 	 */
 	public int getLastGattStatus() {
-		return lastGattStatus;
+		return mLastGattStatus;
+	}
+
+	/**
+	 * Set last gatt status
+	 * @param status status
+	 */
+	private synchronized void setLastGattStatus(int status) {
+		mLastGattStatus = status;
+	}
+
+	/**
+	 * Set last write success
+ 	 * @param success success ?
+	 */
+	private synchronized void setLastWriteSuccess(boolean success) {
+		mLastWriteSuccess = success;
+	}
+
+	/**
+	 * Set write queue
+	 */
+	private synchronized void setWriteQueue(@Nullable Queue<byte[]> queue) {
+		if(mWriteQueue != null) mWriteQueue.clear();
+		mWriteQueue = queue;
+	}
+
+	/**
+	 * Set write callback
+	 * @param callback BleWriteCallback
+	 */
+	private synchronized void setWriteCallback(@Nullable BleWriteCallback callback) {
+		mWriteCallback = callback;
+	}
+
+	/**
+	 * Set read callback
+	 * @param callback BleReadCallback
+	 */
+	private synchronized void setReadCallback(@Nullable BleReadCallback callback) {
+		mReadCallback = callback;
+	}
+
+	/**
+	 * Set the BleGattCallback
+	 * @param calback BleGattCallback
+	 */
+	public synchronized void setGattCallback(@Nullable BleGattCallback calback) {
+		mCallback = calback;
+	}
+
+	/**
+	 * Set BluetoothGatt
+	 * @param gatt BluetoothGatt
+	 */
+	private synchronized void setBluetoothGatt(@Nullable BluetoothGatt gatt) {
+		mBluetoothGatt = gatt;
+	}
+
+	/**
+	 * Set read BluetoothGattCharacteristic
+	 * @param characteristic read BluetoothGattCharacteristic
+	 */
+	private synchronized void setReadCharacteristic(@Nullable BluetoothGattCharacteristic characteristic) {
+		mReadCharacteristic = characteristic;
+	}
+
+	/**
+	 * Set write BluetoothGattCharacteristic
+	 * @param characteristic write BluetoothGattCharacteristic
+	 */
+	private synchronized void setWriteCharacteristic(@Nullable BluetoothGattCharacteristic characteristic) {
+		mWriteCharacteristic = characteristic;
 	}
 
 	/**
@@ -140,25 +212,16 @@ public class DeviceOp {
 	 */
 	private final BluetoothGattCallback mBluetoothGattCallback = new BluetoothGattCallback() {
 		@Override
-		public synchronized void onConnectionStateChange(BluetoothGatt gatt, int status, int newState) {
-			Log.d(LOG, "onConnectionStateChange status" + status + " newState=" + newState);
+		public void onConnectionStateChange(BluetoothGatt gatt, int status, int newState) {
+			Log.d(LOG, "onConnectionStateChange");
 			super.onConnectionStateChange(gatt, status, newState);
-			lastGattStatus = status;
+			setLastGattStatus(status);
 
 			switch (newState) {
 				// Connected
 				case BluetoothProfile.STATE_CONNECTED:
 					// Discover gatt services with 250ms of delay for slow devices
 					mMainHandler.postDelayed(mBluetoothGatt::discoverServices, 250);
-					break;
-
-				// Connecting
-				case BluetoothProfile.STATE_CONNECTING:
-					if(mState == BleGattCallback.CONNECTING) break;
-					mState = BleGattCallback.CONNECTING;
-					mMainHandler.post(() -> {
-						if(mCallback != null) mCallback.onConnecting();
-					});
 					break;
 
 				// Disconnect
@@ -170,57 +233,61 @@ public class DeviceOp {
 		}
 
 		@Override
-		public synchronized void onServicesDiscovered(BluetoothGatt gatt, int status) {
-			Log.d(LOG, "onServicesDiscovered status" + status);
+		public void onServicesDiscovered(BluetoothGatt gatt, int status) {
+			Log.d(LOG, "onServicesDiscovered");
 			super.onServicesDiscovered(gatt, status);
-			lastGattStatus = status;
+			setLastGattStatus(status);
 
-			// Set state connected after services have been discovered
-			if(mState != BleGattCallback.CONNECTED) {
-				mState = BleGattCallback.CONNECTED;
-				mMainHandler.post(() -> {
-					if(mCallback != null) mCallback.onConnectSuccess(gatt.getServices());
-				});
-			}
+			if(mState == BleGattCallback.CONNECTED) return;
+			setState(BleGattCallback.CONNECTED);
+
+			mMainHandler.post(() -> {
+				if(mCallback != null) mCallback.onConnectSuccess(gatt.getServices());
+			});
 		}
 
 		@Override
-		public synchronized void onCharacteristicWrite(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic, int status) {
-			Log.d(LOG, "onCharacteristicWrite characteristic=" + characteristic.getUuid().toString() + " status=" + status);
+		public void onCharacteristicWrite(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic, int status) {
+			Log.d(LOG, "onCharacteristicWrite");
 			super.onCharacteristicWrite(gatt, characteristic, status);
-			lastGattStatus = status;
+			setLastGattStatus(status);
 
 			// Run on the main thread
 			mMainHandler.post(() -> {
 				// if success
 				if(status == BluetoothGatt.GATT_SUCCESS) {
 					// Set last write to success
-					lastWriteSuccess = true;
+					setLastWriteSuccess(true);
 
 					// If write after success
 					if(mWriteAfterSuccess) {
 						// Notify progress
-						if(mWriteCallback != null) mWriteCallback.onWriteProgress(mWriteTotalPkg-mWriteQueue.size(), mWriteTotalPkg);
+						if(mWriteCallback != null && mWriteQueue != null) mWriteCallback.onWriteProgress(mWriteTotalPkg-mWriteQueue.size(), mWriteTotalPkg);
+
 						// Write next packet
 						write();
 					}
 				} else {
 					// Set last write to fail
-					lastWriteSuccess = false;
+					setLastWriteSuccess(false);
 
-					// Clear the write queu
-					if(mWriteQueue != null) mWriteQueue.clear();
+					// Clear the write queue
+					setWriteQueue(null);
+
 					// Notify
-					if(mWriteCallback != null) mWriteCallback.onWriteFailed();
+					if (mWriteCallback != null) {
+						mWriteCallback.onWriteFailed();
+						setWriteCallback(null);
+					}
 				}
 			});
 		}
 
 		@Override
-		public synchronized void onCharacteristicRead(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic, int status) {
-			Log.d(LOG, "onCharacteristicRead characteristic=" + characteristic.getUuid().toString() + " status=" + status);
+		public void onCharacteristicRead(BluetoothGatt gatt, BluetoothGattCharacteristic characteristic, int status) {
+			Log.d(LOG, "onCharacteristicRead");
 			super.onCharacteristicRead(gatt, characteristic, status);
-			lastGattStatus = status;
+			setLastGattStatus(status);
 
 			// Run on the main thread
 			mMainHandler.post(() -> {
@@ -232,9 +299,8 @@ public class DeviceOp {
 				}
 
 				// Clean
-				mReadCallback = null;
-				mReadService = null;
-				mReadCharacteristic = null;
+				setReadCallback(null);
+				setReadCharacteristic(null);
 			});
 		}
 
@@ -283,13 +349,13 @@ public class DeviceOp {
 	 * Connect to the device
 	 * @param callback The callback
 	 */
-	public synchronized void connect(@NonNull BleGattCallback callback) {
+	public void connect(@NonNull BleGattCallback callback) {
 		Log.d(LOG, "connect");
 
 		// Run on the main thread
 		mMainHandler.post(() -> {
 			// Set the callback
-			mCallback = callback;
+			setGattCallback(callback);
 
 			// Check the current state
 			switch (mState) {
@@ -311,40 +377,38 @@ public class DeviceOp {
 			int delayBeforeConnect = 0;
 
 			// Enable bluetooth if not enabled
-			if(!KnBle.isBluetoothEnabled()) {
+			if(!KnBle.getInstance().isBluetoothEnabled()) {
 				Log.d(LOG, "bluetooth is disabled");
-				KnBle.enableBluetooth(true);
+				KnBle.getInstance().enableBluetooth(true);
 				delayBeforeConnect += 5000;
 			}
 
 			// Set state connecting
-			mState = BleGattCallback.CONNECTING;
-
-			// Callback
+			setState(BleGattCallback.CONNECTING);
 			callback.onConnecting();
 
 			// Connecting after the delay
 			mMainHandler.postDelayed(() -> {
-				mBluetoothGatt = null;
+				setBluetoothGatt(null);
 
 				// Always connect with autoConnect==false for better connection speed
 				// Android 6+ (Connect with TRANSPORT_LE)
 				if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-					mBluetoothGatt = mDevice.getDevice().connectGatt(KnBle.getContext(), false, mBluetoothGattCallback, BluetoothDevice.TRANSPORT_LE);
+					setBluetoothGatt(mDevice.getDevice().connectGatt(KnBle.getInstance().getContext(), false, mBluetoothGattCallback, BluetoothDevice.TRANSPORT_LE));
 				}
 				// Android 5.0+ (Connect with reflection method for getting TRANSPORT_LE before Android 6.0)
 				else if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
 					try {
 						Method connectGattMethod = mDevice.getDevice().getClass().getDeclaredMethod("connectGatt", Context.class, boolean.class, BluetoothGattCallback.class, int.class);
 						connectGattMethod.setAccessible(true);
-						mBluetoothGatt = (BluetoothGatt) connectGattMethod.invoke(mDevice.getDevice(), KnBle.getContext(), false, mBluetoothGattCallback, 2);
+						setBluetoothGatt((BluetoothGatt) connectGattMethod.invoke(mDevice.getDevice(), KnBle.getInstance().getContext(), false, mBluetoothGattCallback, 2));
 					} catch (Exception e) {
 						e.printStackTrace();
 					}
 				}
 
 				// If Android 4.4 or If other methods have failed
-				if(mBluetoothGatt == null) mBluetoothGatt =  mDevice.getDevice().connectGatt(KnBle.getContext(), false, mBluetoothGattCallback);
+				if(mBluetoothGatt == null) setBluetoothGatt(mDevice.getDevice().connectGatt(KnBle.getInstance().getContext(), false, mBluetoothGattCallback));
 
 				// All method have failed
 				if(mBluetoothGatt == null) {
@@ -363,7 +427,7 @@ public class DeviceOp {
 	 * @param serviceUUID The service UUID
 	 */
 	public void hasService(@NonNull String serviceUUID, @NonNull BleCheckCallback callback) {
-		Log.d(LOG, "hasService() serviceUUID=" + serviceUUID);
+		Log.d(LOG, "hasService");
 
 		// Run on the main thread
 		mMainHandler.post(() -> {
@@ -373,22 +437,9 @@ public class DeviceOp {
 				return;
 			}
 
-			// Convert to uuid
-			UUID uuid = UUID.fromString(serviceUUID);
-
-			// Check each services
-			for(BluetoothGattService service : mBluetoothGatt.getServices()) {
-				// Check if service UUID is the same
-				if(service.getUuid().equals(uuid)) {
-					Log.d(LOG, "hasService() YES serviceUUID=" + serviceUUID);
-					callback.onResponse(true);
-					return;
-				}
-			}
-
-			// Not found
-			Log.d(LOG, "hasService() NO serviceUUID=" + serviceUUID);
-			callback.onResponse(false);
+			// Get the service
+			BluetoothGattService service = mBluetoothGatt.getService(UUID.fromString(serviceUUID));
+			callback.onResponse(service != null);
 		});
 	}
 
@@ -398,7 +449,8 @@ public class DeviceOp {
 	 * @param characteristicUUID The characteristic UUID
 	 */
 	public void hasCharacteristic(@NonNull String serviceUUID, @NonNull String characteristicUUID, @NonNull BleCheckCallback callback) {
-		Log.d(LOG, "hasCharacteristic() serviceUUID=" + serviceUUID + " characteristicUUID=" + characteristicUUID);
+		Log.d(LOG, "hasCharacteristic");
+
 		// Run on the main thread
 		mMainHandler.post(() -> {
 			// Check if is connected
@@ -410,20 +462,13 @@ public class DeviceOp {
 			// Get the service
 			BluetoothGattService service = mBluetoothGatt.getService(UUID.fromString(serviceUUID));
 			if(service == null) {
-				Log.d(LOG, "hasCharacteristic() NO serviceUUID=" + serviceUUID);
 				callback.onResponse(false);
 				return;
 			}
 
 			// Get the characteristic
 			BluetoothGattCharacteristic characteristic = service.getCharacteristic(UUID.fromString(characteristicUUID));
-			if(characteristic == null) {
-				Log.d(LOG, "hasCharacteristic() YES serviceUUID=" + serviceUUID + " NO characteristicUUID=" + characteristicUUID);
-				callback.onResponse(false);
-			} else {
-				Log.d(LOG, "hasCharacteristic() YES serviceUUID=" + serviceUUID + " YES characteristicUUID=" + characteristicUUID);
-				callback.onResponse(true);
-			}
+			callback.onResponse(characteristic != null);
 		});
 	}
 
@@ -438,7 +483,7 @@ public class DeviceOp {
 	 * @param intervalBetweenTwoPackage Interval between pkg (sendNextWhenLastSuccess == false)
 	 * @param callback The callback
 	 */
-	public synchronized void write(@NonNull String serviceUUID,
+	public void write(@NonNull String serviceUUID,
 								   @NonNull String characteristicUUID,
 								   @NonNull byte[] data,
 								   boolean split,
@@ -447,30 +492,29 @@ public class DeviceOp {
 								   long intervalBetweenTwoPackage,
 								   @NonNull BleWriteCallback callback) {
 
-		Log.d(LOG, "write() serviceUUID=" + serviceUUID + " characteristicUUID=" + characteristicUUID + " datalen=" + data.length
-				+ " split=" + split + " sendNextWhenLastSuccess=" + sendNextWhenLastSuccess + " intervalBetweenTwoPackage=" + intervalBetweenTwoPackage);
+		Log.d(LOG, "write");
 
 		// Run on the main thread
 		mMainHandler.post(() -> {
 			// Check if is connected
 			if(mBluetoothGatt == null) {
-				Log.d(LOG, "write() mBluetoothGatt is null");
+				Log.d(LOG, "write mBluetoothGatt is null");
 				callback.onWriteFailed();
 				return;
 			}
 
 			// Get the service
-			mWriteService = mBluetoothGatt.getService(UUID.fromString(serviceUUID));
-			if(mWriteService == null) {
-				Log.d(LOG, "write() service is null");
+			BluetoothGattService service = mBluetoothGatt.getService(UUID.fromString(serviceUUID));
+			if(service == null) {
+				Log.d(LOG, "write service is null");
 				callback.onWriteFailed();
 				return;
 			}
 
 			// Get the characteristic
-			mWriteCharacteristic = mWriteService.getCharacteristic(UUID.fromString(characteristicUUID));
+			setWriteCharacteristic(service.getCharacteristic(UUID.fromString(characteristicUUID)));
 			if(mWriteCharacteristic == null || (mWriteCharacteristic.getProperties() & (BluetoothGattCharacteristic.PROPERTY_WRITE | BluetoothGattCharacteristic.PROPERTY_WRITE_NO_RESPONSE)) == 0) {
-				Log.d(LOG, "write() characteristic is null");
+				Log.d(LOG, "write characteristic is null");
 				callback.onWriteFailed();
 				return;
 			}
@@ -480,13 +524,15 @@ public class DeviceOp {
 			else mWriteCharacteristic.setWriteType(BluetoothGattCharacteristic.WRITE_TYPE_DEFAULT);
 
 			// Split the data and create the queue
-			mWriteQueue = Utils.splitBytesArray(data, split, spliteSize);
-			mWriteTotalPkg = mWriteQueue.size();
+			setWriteQueue(Utils.splitBytesArray(data, split, spliteSize));
 
-			// Set others vars
-			mWriteAfterSuccess = sendNextWhenLastSuccess;
-			mWriteInterval = intervalBetweenTwoPackage;
-			mWriteCallback = callback;
+			synchronized (DeviceOp.class) {
+				mWriteTotalPkg = mWriteQueue.size();
+				mWriteAfterSuccess = sendNextWhenLastSuccess;
+				mWriteInterval = intervalBetweenTwoPackage;
+			}
+
+			setWriteCallback(callback);
 
 			// Start write
 			write();
@@ -496,26 +542,25 @@ public class DeviceOp {
 	/**
 	 * Write
 	 */
-	private synchronized void write() {
-		Log.d(LOG, "write()");
+	private void write() {
+		Log.d(LOG, "private write");
 
 		// Check if queue is empty
-		if(mWriteQueue.peek() == null) {
+		if(mWriteQueue == null || mWriteQueue.peek() == null) {
 			// If last write is a success
-			if(lastWriteSuccess) mWriteCallback.onWriteSuccess();
-			else Log.d(LOG, "write() lastWriteSuccess==false");
+			if(mLastWriteSuccess) mWriteCallback.onWriteSuccess();
+			else Log.d(LOG, "private write lastWriteSuccess==false");
 
 			// Cleanup
-			mWriteQueue = null;
-			mWriteCharacteristic = null;
-			mWriteService = null;
-			mWriteCallback = null;
+			setWriteCharacteristic(null);
+			setWriteQueue(null);
+			setWriteCallback(null);
 			return;
 		}
 
 		// Check if gatt is not null
 		if(mBluetoothGatt == null) {
-			Log.d(LOG, "write() mBluetoothGatt is null");
+			Log.d(LOG, "private write mBluetoothGatt is null");
 			if(mWriteCallback != null) mWriteCallback.onWriteFailed();
 			return;
 		}
@@ -539,11 +584,11 @@ public class DeviceOp {
 	 * @param connectionPriority priority
 	 */
 	public void requestConnectionPriority(int connectionPriority) {
-		Log.d(LOG, "requestConnectionPriority() connectionPriority=" + connectionPriority);
+		Log.d(LOG, "requestConnectionPriority");
 		if(!isConnected()) return;
 
 		mMainHandler.post(() -> {
-			if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && mBluetoothGatt!= null) mBluetoothGatt.requestConnectionPriority(connectionPriority);
+			if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP && mBluetoothGatt != null) mBluetoothGatt.requestConnectionPriority(connectionPriority);
 		});
 	}
 
@@ -553,36 +598,36 @@ public class DeviceOp {
 	 * @param characteristicUUID The characteristic UUID
 	 * @param callback The callback
 	 */
-	public synchronized void read(@NonNull String serviceUUID, @NonNull String characteristicUUID, @NonNull BleReadCallback callback) {
-		Log.d(LOG, "read() serviceUUID=" + serviceUUID + " characteristicUUID=" + characteristicUUID);
+	public void read(@NonNull String serviceUUID, @NonNull String characteristicUUID, @NonNull BleReadCallback callback) {
+		Log.d(LOG, "read");
 
 		// Run on the main thread
 		mMainHandler.post(() -> {
 			// Check if is connected
 			if(mBluetoothGatt == null) {
-				Log.d(LOG, "read() mBluetoothGatt is null");
+				Log.d(LOG, "read mBluetoothGatt is null");
 				callback.onReadFailed();
 				return;
 			}
 
 			// Get the service
-			mReadService = mBluetoothGatt.getService(UUID.fromString(serviceUUID));
-			if(mReadService == null) {
-				Log.d(LOG, "read() service is null");
+			BluetoothGattService service = mBluetoothGatt.getService(UUID.fromString(serviceUUID));
+			if(service == null) {
+				Log.d(LOG, "read service is null");
 				callback.onReadFailed();
 				return;
 			}
 
 			// Get the characteristic
-			mReadCharacteristic = mReadService.getCharacteristic(UUID.fromString(characteristicUUID));
+			setReadCharacteristic(service.getCharacteristic(UUID.fromString(characteristicUUID)));
 			if(mReadCharacteristic == null || (mReadCharacteristic.getProperties() & BluetoothGattCharacteristic.PROPERTY_READ) == 0) {
-				Log.d(LOG, "read() characteristic is null");
+				Log.d(LOG, "read characteristic is null");
 				callback.onReadFailed();
 				return;
 			}
 
 			// Set the callback
-			mReadCallback = callback;
+			setReadCallback(callback);
 
 			// Read
 			mBluetoothGatt.readCharacteristic(mReadCharacteristic);
@@ -593,9 +638,9 @@ public class DeviceOp {
 	 * Disconnect the device
 	 */
 
-	@SuppressWarnings({"JavaReflectionMemberAccess", "ConstantConditions"})
-	public synchronized void disconnect() {
-		Log.d(LOG, "disconnect()");
+	@SuppressWarnings({"JavaReflectionMemberAccess"})
+	public void disconnect() {
+		Log.d(LOG, "disconnect");
 
 		// Clear the handler
 		mMainHandler.removeCallbacksAndMessages(null);
@@ -606,12 +651,12 @@ public class DeviceOp {
 			if(mState == BleGattCallback.CONNECTING && mCallback != null) mCallback.onConnectFailed();
 
 			// Set state disconnected
-			mState = BleGattCallback.DISCONNECTED;
+			setState(BleGattCallback.DISCONNECTED);
 
 			// Use close instead disconnect to avoid weird behavior (never use disconnect before close)
 			if(mBluetoothGatt != null) {
 				mBluetoothGatt.close();
-				mBluetoothGatt = null;
+				setBluetoothGatt(null);
 			}
 
 			// Unbond the device
@@ -625,21 +670,7 @@ public class DeviceOp {
 			// Callback
 			if(mCallback != null) {
 				mCallback.onDisconnected();
-				mCallback = null;
-			}
-		});
-	}
-
-	/**
-	 * Change the BleGattCallback
-	 * @param calback The callback
-	 */
-	public void setGattCallback(@NonNull BleGattCallback calback) {
-		Log.d(LOG, "setGattCallback()");
-
-		mMainHandler.post(() -> {
-			synchronized (DeviceOp.class) {
-				mCallback = calback;
+				setGattCallback(null);
 			}
 		});
 	}
@@ -649,7 +680,7 @@ public class DeviceOp {
 	 */
 	@SuppressWarnings({"JavaReflectionMemberAccess"})
 	private void clearDeviceCache() {
-		Log.d(LOG, "clearDeviceCache()");
+		Log.d(LOG, "clearDeviceCache");
 
 		mMainHandler.post(() -> {
 			try {